// src/app/models/i-nfl-player-stats.ts
export interface INflPlayerStats {
  player: string;
  team: string;
  position: string;
  attempts: number;
  attemptsPerGame: number;
  yards: number;
  average: string;
  yardsPerGame: number;
  touchdowns: number;
  longestGain: any;
  firstDowns: number;
  firstDownPercentage: number;
  twentyYardRushes: number;
  fourtyYardRushes: number;
  fumbles: number;
}